LJ-backend/backend/requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
certifi==2021.5.30
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.68.0
h11==0.12.0
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
orjson==3.6.3
promise==2.3
pydantic==1.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.26.0
Rx==1.6.1
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
ujson==4.1.0
urllib3==1.26.6
uvicorn[standard]
watchgod==0.7
websockets==8.1
sqlalchemy==1.4.23
postgres==3.0.0
pytz==2021.1
pytest
psycopg2
gunicorn
pathvalidate
google-cloud-storage==1.39.0
google-auth==1.32.0
nats-py==2.1.3